# services/flight-service/src/main/java/com/aeroway/flight/dto/AvailabilityIntegrityResponse.java
package com.aeroway.flight.dto;

public record AvailabilityIntegrityResponse(
        int attempts,
        int successfulReservations,
        int conflicts,
        long duplicateReservationsInDatabase
) {
}