# services/flight-service/src/main/java/com/aeroway/flight/dto/SeatResponse.java
package com.aeroway.flight.dto;

import java.util.UUID;

public record SeatResponse(
        UUID id,
        UUID flightId,
        String seatNumber,
        String cabinClass,
        boolean reserved
) {
}